More fixes for ARM V7

+++ test/unit/algo/reduce.cpp
@@ -76,7 +76,7 @@ void reduce_generic_test_page_ends(eve::as<T> tgt, Init init, Algo alg)
   using card_t  = eve::fixed<4096/ sizeof(e_t)>;
   std::vector<e_t, eve::aligned_allocator<e_t, card_t>> page(card_t::value, e_t{0});
 
-  constexpr int elements_to_test  = std::min(T::size() * 10, 300l);
+  constexpr int elements_to_test  = std::min( int(T::size() * 10), 300);
 
   auto f = page.data();
   auto l = f + elements_to_test;
